chart page form work in progress
