+++ client/package.json
@@ -6,8 +6,9 @@
   "dependencies": {
     "@emotion/react": "^11.8.2",
     "@emotion/styled": "^11.8.1",
-    "@mantine/core": "^4.2.4",
-    "@mantine/hooks": "^4.2.4",
+    "@mantine/core": "^4.2.6",
+    "@mantine/dates": "^4.2.6",
+    "@mantine/hooks": "^4.2.6",
     "@mantine/next": "^4.2.4",
     "@material-ui/core": "^4.12.3",
     "@material-ui/icons": "^4.11.2",
@@ -19,6 +20,7 @@
     "@testing-library/user-event": "^13.5.0",
     "axios": "^0.26.1",
     "chart.js": "^3.7.1",
+    "dayjs": "^1.11.2",
     "faker": "^5.5.3",
     "moment": "^2.29.1",
     "prop-types": "^15.8.1",
